Update script reference and version, add manifest URL
This commit involves several updates in the module.json file. Firstly, the 'digitable.js' script was renamed to 'digitable-initiative-led.js' to better correlate with the module's purpose, hence this change was reflected in the 'scripts' array. The version was also incremented to '1.0.1' indicating a minor change. The 'url' and new 'manifest' fields have been added to provide direct links to the GitHub repository and module manifest file respectively, improving overall module accessibility. The referenced CSS also had its name adjusted to maintain consistency in the files naming convention.