Add download link and update Github Actions release workflow
A download link has been added to module.json to provide a direct link to the versioned module. The Github Actions workflow has been expanded to automate the creation of releases. It now includes steps for installing dependencies, compressing the project into a zip file, creating a release with the established version name, and uploading the compiled project as an asset to the release.